+++ docs/changelog.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- NTVNews.id (`ntvnews`) with bounded Google News sitemap search and latest discovery
+
 ## [1.2.2] - 2026-07-19
 
 ### Added